requirements.txt
discord.py
requests
python-decouple
beautifulsoup4